switchIfEmpty(Single) returns single; remove all changes unrelated to the PR
